third_party/WebKit/LayoutTests/external/wpt/webvtt/webvtt-file-format-parsing/webvtt-cue-text-parsing-rules/tests/timestamps-expected.txt

CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
This is a testharness.js-based test.
FAIL WebVTT cue data parser test timestamps - 54c245f3fbe7a3e25398b13971d44f2bb3a5f947 assert_unreached: got error event Reached unreachable code
FAIL WebVTT cue data parser test timestamps - 5e190a1b4541fcb10e403af111c14ef152fecb0d assert_unreached: got error event Reached unreachable code
FAIL WebVTT cue data parser test timestamps - 92b97d3497836259e0b9305e27f2b91ea1dc9b31 assert_unreached: got error event Reached unreachable code
FAIL WebVTT cue data parser test timestamps - 47fa4306a695161da88533d456ce94829e53b13d assert_unreached: got error event Reached unreachable code
FAIL WebVTT cue data parser test timestamps - c1036a4322c1852e02e5a1843a9a81dfca6d7af3 assert_unreached: got error event Reached unreachable code
FAIL WebVTT cue data parser test timestamps - 70ec34d828ca661a583c651207becb968bb41653 assert_unreached: got error event Reached unreachable code
FAIL WebVTT cue data parser test timestamps - 66ba641ff047a226fa60fe867fd2479d40f3ff0f assert_unreached: got error event Reached unreachable code
FAIL WebVTT cue data parser test timestamps - 398e8da1aaaf392739ca18057fef58bd5333f74d assert_unreached: got error event Reached unreachable code
FAIL WebVTT cue data parser test timestamps - 391fce67644cf4dd9967e1436d1449ef5baf675f assert_unreached: got error event Reached unreachable code
Harness: the test ran to completion.
